Move error handling to error interceptor.
