+++ src/app/presentation/components/member-managment/member-managment.component.ts
@@ -116,28 +116,7 @@ export class MemberManagmentComponent implements OnInit {
         },
 
           err => {
-          let errorMessage: string = '';
-          if(err.error.errors && typeof err.error.errors === 'object'){
-            const errors = err.error.errors;
-
-            for(let key in errors){
-              for(let indexError in errors[key]){
-                errorMessage += errors[key][indexError] + '\n';
-              }
-            }
-
-           this.showAlert(errorMessage);
-
-            return;
-          }
-
-          if(err.error && typeof err.error === 'object'){
-            errorMessage += err.error.error;
-
-            this.showAlert(errorMessage);
-
-            return;
-          }
+            this.showAlert(err);
           }
       );
     }
