+++ src/app/presentation/components/registration/registration.component.ts
@@ -59,29 +59,7 @@ export class RegistrationComponent implements OnInit {
           this.router.navigate(['login']);
         },
         err => {
-          let errorMessage: string = '';
-          if(err.error.errors && typeof err.error.errors === 'object'){
-            const errors = err.error.errors;
-
-            for(let key in errors){
-              for(let indexError in errors[key]){
-                errorMessage += errors[key][indexError] + '\n';
-                console.log(errors[key][indexError]);
-              }
-            }
-
-            this.showAlert(errorMessage);
-
-            return;
-          }
-
-          if(err.error && typeof err.error === 'object'){
-            errorMessage += err.error.error;
-
-            this.showAlert(errorMessage);
-
-            return;
-          }
+          this.showAlert(err);
         }
       )
     }
